FEAT: Error Handler 마무리 정리
