+++ src/main/kotlin/me/snowlight/springkotlincoroutine/config/extension/LocalDateExtension.kt
@@ -1,0 +1,12 @@
+package me.snowlight.springkotlincoroutine.config.extension
+
+import java.time.LocalDate
+import java.time.format.DateTimeFormatter
+
+fun String.toLocalDate(format: String): LocalDate {
+    return LocalDate.parse(this, DateTimeFormatter.ofPattern(format))
+}
+
+fun LocalDate.toString(format: String): String {
+    return this.format(DateTimeFormatter.ofPattern(format))
+}
